Show player's actual avatar on map, refactor image helpers
