+++ Math Go/Controller/MapViewController.swift
@@ -11,19 +11,21 @@ import CoreLocation
 import SwiftUI
 
 /* Notifies the locationManager when it has updated the position of the device
-Source: https://www.raywenderlich.com/764-augmented-reality-ios-tutorial-location-based */
+ Source: https://www.raywenderlich.com/764-augmented-reality-ios-tutorial-location-based */
 class MapViewController: UIViewController {
     
     let appDelegate = UIApplication.shared.delegate as! AppDelegate
-
+    
+    var annotationView: MKAnnotationView? // for Avatar image on map, see extension
+    
     @IBOutlet private var mapView: MKMapView!
-
+    
     @IBAction func inventory(_ sender: Any) {
         let inventoryVC = UIHostingController(rootView: InventoryView().environmentObject(appDelegate.player))
         inventoryVC.modalPresentationStyle = .fullScreen
         self.present(inventoryVC, animated: true)
     }
-
+    
     @IBSegueAction func showCatchView(_ coder: NSCoder) -> UIViewController? {
         // Create MathQuestion and Beastie
         let player = appDelegate.player!
@@ -33,9 +35,9 @@ class MapViewController: UIViewController {
         // Catch Beastie Screen
         return CatchHostingController(coder: coder, beastie: beastie)
     }
-
+    
     fileprivate let locationManager = CLLocationManager()
-
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         // Set initial location to Tampa
@@ -48,14 +50,16 @@ class MapViewController: UIViewController {
         locationManager.requestWhenInUseAuthorization()
         mapView.delegate = self
     }
-
+    
     override func viewWillAppear(_ animated: Bool) {
         // Display Start Screen if no player exists
         if (!appDelegate.playerExists) {
             showStartScreen()
+        } else {
+            annotationView?.image = scaledAvatar()
         }
     }
-
+    
     private func showStartScreen() {
         appDelegate.setPlayerExists(value: true)
         appDelegate.player = Player()
@@ -70,36 +74,35 @@ class MapViewController: UIViewController {
 extension MapViewController: MKMapViewDelegate {
     func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
         let reuseID = "myAnnotationView"
-        var annotationView = mapView.dequeueReusableAnnotationView(withIdentifier: reuseID)
+        annotationView = mapView.dequeueReusableAnnotationView(withIdentifier: reuseID)
         if annotationView == nil {
             annotationView = MKAnnotationView(annotation: annotation, reuseIdentifier: reuseID)
         }
         
-        // TO DO - Get selected avatar to appear here, for now it is just a static avatar
-        let pinImage = UIImage(named: "Avatar5")
-        let targetSize = CGSize(width: 100, height: 100)
-        let scaledImage = pinImage?.scalePreservingAspectRatio(
-            targetSize: targetSize
-        )
-        annotationView?.image = scaledImage
-
+        annotationView?.image = scaledAvatar()
+        
         return annotationView
+    }
+    
+    // Return scaled Avatar image
+    func scaledAvatar() -> UIImage {
+        return UIImage.scaleImage(imageName: appDelegate.player.avatar, size: 100)
     }
 }
 
 /* Displays region at specified zoom level
-Source: https://www.raywenderlich.com/7738344-mapkit-tutorial-getting-started */
+ Source: https://www.raywenderlich.com/7738344-mapkit-tutorial-getting-started */
 private extension MKMapView {
-  func centerToLocation(
-    _ location: CLLocation,
-    regionRadius: CLLocationDistance = 1000
-  ) {
-    let coordinateRegion = MKCoordinateRegion(
-      center: location.coordinate,
-      latitudinalMeters: regionRadius,
-      longitudinalMeters: regionRadius)
-    setRegion(coordinateRegion, animated: true)
-  }
+    func centerToLocation(
+        _ location: CLLocation,
+        regionRadius: CLLocationDistance = 1000
+    ) {
+        let coordinateRegion = MKCoordinateRegion(
+            center: location.coordinate,
+            latitudinalMeters: regionRadius,
+            longitudinalMeters: regionRadius)
+        setRegion(coordinateRegion, animated: true)
+    }
 }
 
 /* Delegate method that gets the current location of the user */
@@ -111,38 +114,5 @@ extension MapViewController: CLLocationManagerDelegate {
         
         mapView.setRegion(region, animated: true)
         mapView.showsUserLocation = true
-    }
-}
-
-
-/* Resizes UIImage without scretching
- Source: https://www.advancedswift.com/resize-uiimage-no-stretching-swift/ */
-extension UIImage {
-    func scalePreservingAspectRatio(targetSize: CGSize) -> UIImage {
-        // Determine the scale factor that preserves aspect ratio
-        let widthRatio = targetSize.width / size.width
-        let heightRatio = targetSize.height / size.height
-        
-        let scaleFactor = min(widthRatio, heightRatio)
-        
-        // Compute the new image size that preserves aspect ratio
-        let scaledImageSize = CGSize(
-            width: size.width * scaleFactor,
-            height: size.height * scaleFactor
-        )
-
-        // Draw and return the resized UIImage
-        let renderer = UIGraphicsImageRenderer(
-            size: scaledImageSize
-        )
-
-        let scaledImage = renderer.image { _ in
-            self.draw(in: CGRect(
-                origin: .zero,
-                size: scaledImageSize
-            ))
-        }
-        
-        return scaledImage
     }
 }
