+++ Math Go/Styles.swift
@@ -1,5 +1,6 @@
 import Foundation
 import SwiftUI
+import UIKit
 
 /* Global Styles */
 
@@ -58,5 +59,45 @@ extension Image {
         self
             .resizable()
             .scaledToFit()
+    }
+}
+
+/* Resizes UIImage without scretching
+ Source: https://www.advancedswift.com/resize-uiimage-no-stretching-swift/ */
+extension UIImage {
+    func scalePreservingAspectRatio(targetSize: CGSize) -> UIImage {
+        // Determine the scale factor that preserves aspect ratio
+        let widthRatio = targetSize.width / size.width
+        let heightRatio = targetSize.height / size.height
+        
+        let scaleFactor = min(widthRatio, heightRatio)
+        
+        // Compute the new image size that preserves aspect ratio
+        let scaledImageSize = CGSize(
+            width: size.width * scaleFactor,
+            height: size.height * scaleFactor
+        )
+        
+        // Draw and return the resized UIImage
+        let renderer = UIGraphicsImageRenderer(
+            size: scaledImageSize
+        )
+        
+        let scaledImage = renderer.image { _ in
+            self.draw(in: CGRect(
+                origin: .zero,
+                size: scaledImageSize
+            ))
+        }
+        
+        return scaledImage
+    }
+    
+    // Helper method to scale a UIImage
+    static func scaleImage(imageName: String, size: Int) -> UIImage {
+        let pinImage = UIImage(named: imageName)
+        let targetSize = CGSize(width: size, height: size)
+        let scaledImage = pinImage?.scalePreservingAspectRatio(targetSize: targetSize)
+        return scaledImage ?? UIImage()
     }
 }
